TODO holiday api

+++ internal/app/user/services.go
@@ -21,6 +21,7 @@ type (
 		FindByEmail(ctx context.Context, email string) (*types.User, error)
 		FindAllDev(ctx context.Context, createrID string) ([]*types.User, error)
 		Delete(cxt context.Context, id string) error
+		FindByID(ctx context.Context, UserID string) (*types.User, error)
 	}
 
 	PolicyService interface {
@@ -71,6 +72,7 @@ func (s *Service) Register(ctx context.Context, req *types.RegisterRequest) (*ty
 		Email:     req.Email,
 		Role:      types.PM,
 		UserID:    uuid.New(),
+		ProjectID: []string{},
 	}
 
 	if _, err := s.repo.Create(ctx, user); err != nil {
@@ -92,7 +94,6 @@ func (s *Service) CreateDev(ctx context.Context, req *types.RegisterRequest) (*t
 	}
 
 	existingUser, err := s.repo.FindByEmail(ctx, req.Email)
-
 	if err != nil && !db.IsErrNotFound(err) {
 		logrus.WithContext(ctx).Errorf("failed to check existing user by email, err: %v", err)
 		return nil, fmt.Errorf("failed to check existing user by email: %w", err)
@@ -178,6 +179,14 @@ func (s *Service) FindAllDev(ctx context.Context) ([]*types.UserInfo, error) {
 func (s *Service) Delete(ctx context.Context, id string) error {
 	if err := s.policy.Validate(ctx, types.PolicyObjectDeleteDev, types.PolicyActionDevDelete); err != nil {
 		return err
+	}
+	user, err := s.repo.FindByID(ctx, id)
+	if err != nil {
+		logrus.WithContext(ctx).Errorf("failed to check existing user by ID, err: %v", err)
+		return status.Gen().BadRequest
+	}
+	if user.Role == types.PM {
+		return status.Sercurity().InvalidAction
 	}
 	return s.repo.Delete(ctx, id)
 }
